+++ Simp1e.Editor/xmake.lua
@@ -3,7 +3,7 @@ add_languages("c++20")
 target("Simp1e.Editor")
     add_qt()
     set_kind("binary")
-    add_files("Editor.cpp", "../Resources/Editor.qrc", "include/**/Q_*.h")
+    add_files("Editor.cpp", "../Resources/Editor.qrc", "include/**/*_Q.h")
     add_includedirs("include")
     add_headerfiles("include/(**).h")
     add_deps("Simp1e.Data")
